[Sibz.UXML List] Refactor part 2
