+++ com.sibz.uxml-list/Editor/ListElement.Interfaces.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 78c60f0e51ab145469ab3c3bbc1d934f
+guid: 6293679f8690afa4f8520457b66aee2b
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
