--- com.sibz.uxml-list/Editor/ListElementType.cs
@@ -1,19 +1,0 @@
-﻿namespace Sibz.UXMLList
-{
-    public enum ListElementType
-    {
-        HeaderSection,
-        Label,
-        AddButton,
-        DeleteAllButton,
-        DeleteAllConfirmSection,
-        DeleteAllConfirmLabel,
-        DeleteAllConfirmButton,
-        DeleteAllCancelButton,
-        ListItemSection,
-        ListItem,
-        MoveUpButton,
-        MoveDownButton,
-        DeleteItemButton
-    }
-}
