--- com.sibz.uxml-list/Editor/ListItemElementType.cs
@@ -1,9 +1,0 @@
-﻿namespace Sibz.UXMLList
-{
-    public enum ListItemElementType
-    {
-        MoveUp,
-        MoveDown,
-        Delete
-    }
-}
